% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanSSM-package.R
\docType{package}
\name{kalmanSSM-package}
\alias{kalmanSSM-package}
\alias{_PACKAGE}
\title{kalmanSSM: Kalman Filter}
\description{
Functions related to Kalman filters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ijapesigan/kalmanSSM}
  \item \url{https://ijapesigan.github.io/kalmanSSM/}
  \item Report bugs at \url{https://github.com/ijapesigan/kalmanSSM/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.ijapesigan@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
